extract test_blockchain_query_by_number_gets_canonical to easily test multiple impls
