+++ core/client/db/src/lib.rs
@@ -1089,4 +1089,19 @@ mod tests {
 
 		test_client::trait_tests::test_leaves_for_backend(backend);
 	}
+
+	#[test]
+	fn test_blockchain_query_by_number_gets_canonical() {
+		let database_settings = DatabaseSettings {
+			cache_size: None,
+			// TODO [snd] tmp dir
+			path: "tmp".into(),
+			// TODO [snd] investigate
+			pruning: PruningMode::ArchiveAll,
+		};
+
+		let backend: Arc<Backend<test_client::runtime::Block>> = Arc::new(Backend::new(database_settings, FINALIZATION_WINDOW).unwrap());
+
+		test_client::trait_tests::test_blockchain_query_by_number_gets_canonical(backend);
+	}
 }
